refactor(mobile): 최근 역 검색 목록 분리
